+++ heating_assistant/custom_components/heating_assistant/const.py
@@ -2,7 +2,7 @@
 
 DOMAIN = "heating_assistant"
 NAME = "Heating Assistant"
-VERSION = "2026.09.22"
+VERSION = "2026.09.23"
 
 CONF_INSTANCE_ID = "instance_id"
 DEFAULT_INSTANCE_ID = "default"